Moved classes for day 3
